Moved (I)Repos to correct folders and updated references
