+++ StonksBackend/Domain/DataStores/IDataRepo.cs
@@ -1,0 +1,12 @@
+
+namespace StonksBackend.Domain
+{
+    public interface IDataRepo<T>
+    {
+
+        Task<T> GetByIDAsync(String id);
+        Task<T> AddEntityAsync(T entity);
+        Task<T> DeleteAsync(T entity);
+
+    }
+}
