+++ StonksBackend/Domain/DataStores/IUserRepo.cs
@@ -2,14 +2,7 @@ using StonksBackend.Domain.Entities;
 
 namespace StonksBackend.Domain.DataStores{
 
-    public interface IUserRepo{
-
-        public Task GetUserByIDAsync(String id);
-
-        public Task CreateUserAsync(User user);
-
-        public Task UpdateUserAsync(User user);
-
+    public interface IUserRepo : IDataRepo<User>{
 
     }
 }
