--- StonksBackend/Domain/IDataRepo.cs
@@ -1,13 +1,0 @@
-
-namespace StonksBackend.Domain
-{
-    public interface IDataRepo<T>
-    {
-
-        Task<T> getByIDAsync(String id);
-        Task<T> addAsync(T entity);
-        Task<T> updateAsync(T entity);
-        Task<T> deleteAsync(T entity);
-
-    }
-}
